Reorganize tests, put in submodules.
